extract updateLicenseDump to separate module and promisify it
